refactor: Create driver.dto package
Also, DriverDataPage is a DTO, so it's renamed to reflect it.
